test code to test asm
