+++ test/Sentry.Tests/Protocol/BreadcrumbTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Sentry.Internal;
 using Sentry.Protocol;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 namespace Sentry.Tests.Protocol
