+++ test/Sentry.Tests/Protocol/Context/BrowserTests.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using FluentAssertions;
 using Sentry.Internal;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable once CheckNamespace
