+++ test/Sentry.Tests/Protocol/Context/DeviceTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable once CheckNamespace
