+++ test/Sentry.Tests/Protocol/Context/RuntimeTests.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable once CheckNamespace
