+++ test/Sentry.Tests/Protocol/Context/TraceTests.cs
@@ -1,4 +1,5 @@
 ﻿using Sentry.Protocol;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 namespace Sentry.Tests.Protocol.Context
