+++ test/Sentry.Tests/Protocol/Exceptions/MechanismTests.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable once CheckNamespace
