+++ test/Sentry.Tests/Protocol/SentryMessageTests.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Sentry.Protocol;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 namespace Sentry.Tests.Protocol
