+++ test/Sentry.Tests/Protocol/TransactionTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using NSubstitute;
 using Sentry.Extensibility;
 using Sentry.Internal;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 namespace Sentry.Tests.Protocol
