+++ test/Sentry.Tests/Protocol/UserFeedbackTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using FluentAssertions;
 using Sentry.Protocol;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 namespace Sentry.Tests.Protocol
